Inline registration into UseStrongTypes; rename to Startup.cs
Drop the public StrongTypesWpf.Register() entry point. The
TypeDescriptionProvider install now lives directly inside
UseStrongTypes(this Application), with the idempotency guard
on the static fields. The class becomes ApplicationStartupExtensions
in Startup.cs — file name and class name describe the role rather
than restating the namespace.

The test ModuleInitializer constructs an Application on an STA
thread and calls UseStrongTypes; that's the same path real WPF
apps take, so the tests no longer rely on a back-door API.

Trim the package readme line about non-WPF callers — the only
real caller was tests, which now go through UseStrongTypes too.
