+++ src/StrongTypes.Wpf.Tests/TestSetup.cs
@@ -1,11 +1,12 @@
 #nullable enable
 
 using System.Runtime.CompilerServices;
+using System.Windows;
 
 namespace StrongTypes.Wpf.Tests;
 
 internal static class TestSetup
 {
     [ModuleInitializer]
-    internal static void Init() => StrongTypesWpf.Register();
+    internal static void Init() => StaThread.Run(() => new Application().UseStrongTypes());
 }
